VRF v2.5

+++ script/DeployRaffle.s.sol
@@ -14,7 +14,7 @@ contract DeployRaffle is Script {
             uint256 interval,
             address vrfCoordinator,
             bytes32 keyHash,
-            uint64 subscriptionId,
+            uint256 subscriptionId,
             uint32 callbackGasLimit,
             address link,
             uint256 deployerKey
